было как то сложно с си_пи_пи файлами и я их убрал так выходила с ними ошибка линковщика 2019
